requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
importlib-metadata==3.4.0
Markdown==3.3.3
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
